+++ packages/core-js-compat/src/data.mjs
@@ -1985,7 +1985,11 @@ export const data = {
   },
   'esnext.iterator.to-async': {
   },
+  'esnext.json.is-raw-json': {
+  },
   'esnext.json.parse': {
+  },
+  'esnext.json.raw-json': {
   },
   'esnext.map.delete-all': {
   },