+++ packages/core-js/internals/native-raw-json.js
@@ -1,0 +1,8 @@
+/* eslint-disable es/no-json -- safe */
+var fails = require('../internals/fails');
+
+module.exports = !fails(function () {
+  var unsafeInt = '9007199254740993';
+  var raw = JSON.rawJSON(unsafeInt);
+  return !JSON.isRawJSON(raw) || JSON.stringify(raw) !== unsafeInt;
+});