+++ tests/compat/tests.js
@@ -222,6 +222,12 @@ function createStringTrimMethodTest(METHOD_NAME) {
   };
 }
 
+function NATIVE_RAW_JSON() {
+  var unsafeInt = '9007199254740993';
+  var raw = JSON.rawJSON(unsafeInt);
+  return JSON.isRawJSON(raw) && JSON.stringify(raw) === unsafeInt;
+}
+
 function IMMEDIATE() {
   return setImmediate && clearImmediate && !(IS_BUN && (function () {
     var version = global.Bun.version.split('.');
@@ -1579,6 +1585,7 @@ GLOBAL.tests = {
   'esnext.iterator.to-async': function () {
     return Iterator.prototype.toAsync;
   },
+  'esnext.json.is-raw-json': NATIVE_RAW_JSON,
   'esnext.json.parse': function () {
     var unsafeInt = '9007199254740993';
     var source;
@@ -1587,6 +1594,7 @@ GLOBAL.tests = {
     });
     return source === unsafeInt;
   },
+  'esnext.json.raw-json': NATIVE_RAW_JSON,
   'esnext.map.delete-all': function () {
     return Map.prototype.deleteAll;
   },