+++ CHANGELOG.md
@@ -11,6 +11,12 @@
     - `%TypedArray%.prototype.with`
   - Moved to stable ES, [January 2023 TC39 meeting](https://github.com/babel/proposals/issues/86#issuecomment-1409261397)
   - Added `es.` namespace modules, `/es/` and `/stable/` namespaces entries
+- Added [`JSON.parse` source text access Stage 3 proposal](https://github.com/tc39/proposal-json-parse-with-source)
+  - Methods:
+    - `JSON.parse` patched for support `source` in `reviver` function arguments
+    - `JSON.rawJSON`
+    - `JSON.isRawJSON`
+    - `JSON.stringify` patched for support `JSON.rawJSON`
 - [Iterator Helpers](https://github.com/tc39/proposal-iterator-helpers) proposal:
   - Splitted into 2 ([sync](https://github.com/tc39/proposal-iterator-helpers) and [async](https://github.com/tc39/proposal-async-iterator-helpers)) proposals, async version moved back to Stage 2, [January 2023 TC39 meeting](https://github.com/babel/proposals/issues/86#issuecomment-1410926068)
   - Allowed interleaved mapping in `AsyncIterator` helpers, [proposal-iterator-helpers/262](https://github.com/tc39/proposal-iterator-helpers/pull/262)